@@ -446,7 +446,11 @@ var inChrome = typeof Components != "undefined" && "utils" in Components;
   // We can alias `subarray` to `slice` when the latter is not available, because
   // they're semantically identical.
   if (!Uint8Array.prototype.slice) {
+    /* eslint-disable */
+    // Eslint disabled for no-extend-native; Specific override needed for Firefox 37
+    // and earlier, also for other browsers.
     Uint8Array.prototype.slice = Uint8Array.prototype.subarray;
+    /* eslint-enable */
   }
 
   /**
@@ -736,6 +740,34 @@ var inChrome = typeof Components != "undefined" && "utils" in Components;
     return obj;
   }
 
+  /**
+   * Truncate a string if it exceeds the length as defined in `maxLen`, which
+   * is defined as '72' characters by default. If the string needs trimming,
+   * it'll be suffixed with the unicode ellipsis char, \u2026.
+   *
+   * @param  {String} str    The string to truncate, if needed.
+   * @param  {Number} maxLen Maximum number of characters that the string is
+   *                         allowed to contain. Optional, defaults to 72.
+   * @return {String} Truncated version of `str`.
+   */
+  function truncate(str, maxLen) {
+    maxLen = maxLen || 72;
+
+    if (str.length > maxLen) {
+      var substring = str.substr(0, maxLen);
+      // XXX Due to the fact that we have two different l10n libraries.
+      var direction = mozL10n.getDirection ? mozL10n.getDirection() :
+                      mozL10n.language.direction;
+      if (direction === "rtl") {
+        return "…" + substring;
+      }
+
+      return substring + "…";
+    }
+
+    return str;
+  }
+
   this.utils = {
     CALL_TYPES: CALL_TYPES,
     FAILURE_DETAILS: FAILURE_DETAILS,
@@ -764,6 +796,7 @@ var inChrome = typeof Components != "undefined" && "utils" in Components;
     strToUint8Array: strToUint8Array,
     Uint8ArrayToStr: Uint8ArrayToStr,
     objectDiff: objectDiff,
-    stripFalsyValues: stripFalsyValues
+    stripFalsyValues: stripFalsyValues,
+    truncate: truncate
   };
 }).call(inChrome ? this : loop.shared);
