@@ -14,7 +14,13 @@
   navigator.mozL10n = document.mozL10n = {
     initialize: function(){},
 
-    getDirection: function(){},
+    getDirection: function(){
+      if (document.location.search === "?rtl=1") {
+        return "rtl";
+      }
+
+      return "ltr";
+    },
 
     get: function(stringId, vars) {
 
