@@ -15,6 +15,7 @@
 
   // 1. Desktop components
   // 1.1 Panel
+  var AvailabilityDropdown = loop.panel.AvailabilityDropdown;
   var PanelView = loop.panel.PanelView;
   var SignInRequestView = loop.panel.SignInRequestView;
   // 1.2. Conversation Window
@@ -438,7 +439,17 @@
 
   // Local mocks
 
-  var mockMozLoopRooms = _.extend({}, navigator.mozLoop);
+  var mockMozLoopLoggedIn = _.cloneDeep(navigator.mozLoop);
+  mockMozLoopLoggedIn.userProfile = {
+    email: "text@example.com",
+    uid: "0354b278a381d3cb408bb46ffc01266"
+  };
+
+  var mockMozLoopLoggedInLongEmail = _.cloneDeep(navigator.mozLoop);
+  mockMozLoopLoggedInLongEmail.userProfile = {
+    email: "reallyreallylongtext@example.com",
+    uid: "0354b278a381d3cb408bb46ffc01266"
+  };
 
   var mockContact = {
     name: ["Mr Smith"],
@@ -492,7 +503,8 @@
       "10x10": ["close", "close-active", "close-disabled", "dropdown",
         "dropdown-white", "dropdown-active", "dropdown-disabled", "edit",
         "edit-active", "edit-disabled", "edit-white", "expand", "expand-active",
-        "expand-disabled", "minimize", "minimize-active", "minimize-disabled"
+        "expand-disabled", "minimize", "minimize-active", "minimize-disabled",
+        "settings-cog"
       ],
       "14x14": ["audio", "audio-active", "audio-disabled", "facemute",
         "facemute-active", "facemute-disabled", "hangup", "hangup-active",
@@ -509,7 +521,8 @@
         "precall", "precall-hover", "precall-active", "screen-white", "screenmute-white",
         "settings", "settings-hover", "settings-active", "share-darkgrey", "tag",
         "tag-hover", "tag-active", "trash", "unblock", "unblock-hover", "unblock-active",
-        "video", "video-hover", "video-active", "tour"
+        "video", "video-hover", "video-active", "tour", "status-available",
+        "status-unavailable"
       ]
     },
 
@@ -580,6 +593,7 @@
         React.PropTypes.element,
         React.PropTypes.arrayOf(React.PropTypes.element)
       ]).isRequired,
+      cssClass: React.PropTypes.string,
       dashed: React.PropTypes.bool,
       style: React.PropTypes.object,
       summary: React.PropTypes.string.isRequired
@@ -591,8 +605,14 @@
 
     render: function() {
       var cx = React.addons.classSet;
+      var extraCSSClass = {
+        "example": true
+      };
+      if (this.props.cssClass) {
+        extraCSSClass[this.props.cssClass] = true;
+      }
       return (
-        <div className="example">
+        <div className={cx(extraCSSClass)}>
           <h3 id={this.makeId()}>
             {this.props.summary}
             <a href={this.makeId("#")}>&nbsp;¶</a>
@@ -693,25 +713,31 @@
               <strong>Note:</strong> 332px wide.
             </p>
             <Example dashed={true} style={{width: "332px"}} summary="Re-sign-in view">
-              <SignInRequestView mozLoop={mockMozLoopRooms} />
+              <SignInRequestView mozLoop={mockMozLoopLoggedIn} />
             </Example>
             <Example dashed={true} style={{width: "332px"}} summary="Room list tab">
               <PanelView client={mockClient}
                          dispatcher={dispatcher}
-                         mozLoop={mockMozLoopRooms}
+                         mozLoop={mockMozLoopLoggedIn}
                          notifications={notifications}
                          roomStore={roomStore}
-                         selectedTab="rooms"
-                         userProfile={{email: "test@example.com"}} />
+                         selectedTab="rooms" />
             </Example>
             <Example dashed={true} style={{width: "332px"}} summary="Contact list tab">
               <PanelView client={mockClient}
                          dispatcher={dispatcher}
-                         mozLoop={mockMozLoopRooms}
+                         mozLoop={mockMozLoopLoggedIn}
                          notifications={notifications}
                          roomStore={roomStore}
-                         selectedTab="contacts"
-                         userProfile={{email: "test@example.com"}} />
+                         selectedTab="contacts" />
+            </Example>
+            <Example dashed={true} style={{width: "332px"}} summary="Contact list tab long email">
+              <PanelView client={mockClient}
+                         dispatcher={dispatcher}
+                         mozLoop={mockMozLoopLoggedInLongEmail}
+                         notifications={notifications}
+                         roomStore={roomStore}
+                         selectedTab="contacts" />
             </Example>
             <Example dashed={true} style={{width: "332px"}} summary="Error Notification">
               <PanelView client={mockClient}
@@ -723,26 +749,38 @@
             <Example dashed={true} style={{width: "332px"}} summary="Error Notification - authenticated">
               <PanelView client={mockClient}
                          dispatcher={dispatcher}
-                         mozLoop={navigator.mozLoop}
+                         mozLoop={mockMozLoopLoggedIn}
                          notifications={errNotifications}
-                         roomStore={roomStore}
-                         userProfile={{email: "test@example.com"}} />
+                         roomStore={roomStore} />
             </Example>
             <Example dashed={true} style={{width: "332px"}} summary="Contact import success">
               <PanelView dispatcher={dispatcher}
-                         mozLoop={mockMozLoopRooms}
+                         mozLoop={mockMozLoopLoggedIn}
                          notifications={new loop.shared.models.NotificationCollection([{level: "success", message: "Import success"}])}
                          roomStore={roomStore}
-                         selectedTab="contacts"
-                         userProfile={{email: "test@example.com"}} />
+                         selectedTab="contacts" />
             </Example>
             <Example dashed={true} style={{width: "332px"}} summary="Contact import error">
               <PanelView dispatcher={dispatcher}
-                         mozLoop={mockMozLoopRooms}
+                         mozLoop={mockMozLoopLoggedIn}
                          notifications={new loop.shared.models.NotificationCollection([{level: "error", message: "Import error"}])}
                          roomStore={roomStore}
-                         selectedTab="contacts"
-                         userProfile={{email: "test@example.com"}} />
+                         selectedTab="contacts" />
+            </Example>
+          </Section>
+
+          <Section name="Availability Dropdown">
+            <p className="note">
+              <strong>Note:</strong> 332px wide.
+            </p>
+            <Example dashed={true} style={{width: "332px", height: "200px"}}
+                     summary="AvailabilityDropdown">
+              <AvailabilityDropdown />
+            </Example>
+            <Example cssClass="force-menu-show" dashed={true}
+                     style={{width: "332px", height: "200px"}}
+                     summary="AvailabilityDropdown Expanded">
+              <AvailabilityDropdown />
             </Example>
           </Section>
 
@@ -753,7 +791,7 @@
                 <AcceptCallView callType={CALL_TYPES.AUDIO_VIDEO}
                                 callerId="Mr Smith"
                                 dispatcher={dispatcher}
-                                mozLoop={mockMozLoopRooms} />
+                                mozLoop={mockMozLoopLoggedIn} />
               </div>
             </Example>
 
@@ -763,7 +801,7 @@
                 <AcceptCallView callType={CALL_TYPES.AUDIO_ONLY}
                                 callerId="Mr Smith"
                                 dispatcher={dispatcher}
-                                mozLoop={mockMozLoopRooms} />
+                                mozLoop={mockMozLoopLoggedIn} />
               </div>
             </Example>
           </Section>
@@ -775,7 +813,7 @@
                 <AcceptCallView callType={CALL_TYPES.AUDIO_VIDEO}
                                 callerId="Mr Smith"
                                 dispatcher={dispatcher}
-                                mozLoop={mockMozLoopRooms}
+                                mozLoop={mockMozLoopLoggedIn}
                                 showMenu={true} />
               </div>
             </Example>
