@@ -5,14 +5,14 @@
  * Tests that the memory call tree view renders content after recording.
  */
 function* spawnTest() {
-  let { panel } = yield initPerformance(SIMPLE_URL);
-  let { EVENTS, DetailsView, MemoryCallTreeView } = panel.panelWin;
+  let { panel } = yield initPerformance(ALLOCS_URL);
+  let { EVENTS, $$, PerformanceController, DetailsView, MemoryCallTreeView } = panel.panelWin;
 
   // Enable memory to test.
   Services.prefs.setBoolPref(ALLOCATIONS_PREF, true);
 
   yield startRecording(panel);
-  yield busyWait(100);
+  yield waitUntil(() => PerformanceController.getCurrentRecording().getAllocations().timestamps.length);
   yield stopRecording(panel);
 
   let rendered = once(MemoryCallTreeView, EVENTS.MEMORY_CALL_TREE_RENDERED);
@@ -21,6 +21,8 @@ function* spawnTest() {
   yield rendered;
 
   ok(true, "MemoryCallTreeView rendered after recording is stopped.");
+
+  ok($$("#memory-calltree-view .call-tree-item").length, "there are several allocations rendered.");
 
   yield startRecording(panel);
   yield busyWait(100);
