@@ -157,7 +157,7 @@ public class testAddSearchEngine extends AboutHomeTest {
                 }
 
                 final int actualCount = searchResultList.getAdapter().getCount()
-                        + searchEngineBar.getAdapter().getCount()
+                        + searchEngineBar.getAdapter().getItemCount()
                         - 1; // Subtract one for the search engine bar label (Bug 1172071)
 
                 return (actualCount == expectedCount);
